Comment out file operations in forktest related to memory leak.
When processes open and close files in forktest, the kernel heap
usage does not return to the same value, indicating a memory leak.
Cause is currently unknown, but if we comment out those operations
which are unnecessary for the test, there is no leak.  But it appears
there is a race condition with multiple processes related to file
operations that can result in a memory leak.