Remove lock_do_i_hold asserts and kprintf.
lock_do_i_hold evaluates to true unconditionally during early boot
phase, so is not valid to use.
kprintf may be causing deadlock problems because it uses shared I/O.
If a thread is blocking I/O and waiting for coremap_lock, then
the thread holding the coremap_lock will deadlock if it waits for I/O.
So to be safe, don't use kprintf inside the coremap functions.