Fix corrupted swapmap.
handle_write_fault is responsible for creating new page copies
on write.  The new copy is created in memory but swap is not
copied.  There was a bug because the new copy pte pointed to
the same swap block as the original making the copy still linked
to the original through the swap block.  Fixed so new copy
has VM_PTE_BACKED bit cleared.

Add some diagnostic functions, such as dump_swapmap.